Move main ping tests to own file
